feat(api): Add polling for video translation updates
This commit adds polling functionality to the API for retrieving video translation updates. The polling interval is set to 5000 milliseconds. The code initiates a request to get videos by vid translation and then periodically refetches the data using the polling interval. If an error occurs during polling, it is logged and the polling is stopped. Once the data is retrieved, it is used to update the cached data. If the retrieved data indicates that the processing is complete, the polling is stopped. If no data is retrieved, the polling is stopped as well.